shared types
