+++ shared/types.ts
@@ -1,0 +1,13 @@
+export type DomainAssessment = {
+  domain: string;
+  isPossible: boolean;
+  isAvailable: boolean;
+  isCheap: boolean;
+  evoc: number;
+  brev: number;
+  grep: number;
+  goog: number;
+  pron: number;
+  spel: number;
+  verb: number;
+};
